+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.2.0
 * feat: Print title and data on the same line to avoid log order errors caused by processing large JSON
+* feat: Add DateTime.now() to Req/Resp/Error header line
+* refactor: use StringBuffer collect all log in one req/resp
 
 ## 0.1.0
 * feat: add emoji header